+++ CHANGELOG.md
@@ -37,6 +37,7 @@
    - Unmanaged Device
    - Uplink Set
    - Volume
+   - Volume attachment
    - Volume template
  2. New features added:
    - Drive Enclosures